Revert "✨ Added email design customization"
This reverts commit 4639514a3b3f3976ff867048daac4183e5c6d6f3.
