+++ ghost/admin/app/components/modals/newsletters/edit.hbs
@@ -8,7 +8,11 @@
                 <div class="modal-fullsettings-body-labs">
                     <div class="modal-fullsettings-top">
                         <Modals::Newsletters::Edit::Settings @newsletter={{@data.newsletter}} @openSection={{this.openSection}} @toggleSection={{this.toggleSection}} />
-                        <Modals::Newsletters::Edit::Design @newsletter={{@data.newsletter}} @openSection={{this.openSection}} @toggleSection={{this.toggleSection}} />
+                        {{#if (feature 'makingItRain')}}
+                            <Modals::Newsletters::Edit::DesignLabs @newsletter={{@data.newsletter}} @openSection={{this.openSection}} @toggleSection={{this.toggleSection}} />
+                        {{else}}
+                            <Modals::Newsletters::Edit::Design @newsletter={{@data.newsletter}} @openSection={{this.openSection}} @toggleSection={{this.toggleSection}} />
+                        {{/if}}
                     </div>
                     <div class="modal-fullsettings-bottom">
                         <div class="modal-fullsettings-footer">
@@ -57,7 +61,11 @@
                         data-test-button="save-newsletter"
                     />
                 </div>
-                <Modals::Newsletters::Edit::Preview @newsletter={{@data.newsletter}} />
+                {{#if (feature 'makingItRain')}}
+                    <Modals::Newsletters::Edit::PreviewLabs @newsletter={{@data.newsletter}} />
+                {{else}}
+                    <Modals::Newsletters::Edit::Preview @newsletter={{@data.newsletter}} />
+                {{/if}}
             </div>
         </div>
     </div>
