+++ ghost/admin/app/components/modals/newsletters/edit/design-labs.hbs
@@ -60,7 +60,7 @@
                         {{#if this.settings.icon}}
                             <GhFormGroup class="gh-stack-item row">
                                 <label class="modal-fullsettings-title {{unless this.settings.icon "disabled"}}">Publication icon</label>
-                                <div class="for-switch small {{unless this.settings.icon "disabled"}}">
+                                <div class="for-switch x-small {{unless this.settings.icon "disabled"}}">
                                     <label class="switch" for="show-header">
                                         <input
                                             type="checkbox"
@@ -78,7 +78,7 @@
 
                         <GhFormGroup class="gh-stack-item row">
                             <label class="modal-fullsettings-title">Publication title</label>
-                            <div class="for-switch small">
+                            <div class="for-switch x-small">
                                 <label class="switch" for="show-title" data-test-toggle="showHeaderTitle">
                                     <input
                                         type="checkbox"
@@ -93,7 +93,7 @@
                         </GhFormGroup>
                         <GhFormGroup class="gh-stack-item row">
                             <label class="modal-fullsettings-title">Newsletter name</label>
-                            <div class="for-switch small">
+                            <div class="for-switch x-small">
                                 <label class="switch" for="show-header-name" data-test-toggle="showHeaderName">
                                     <input
                                         type="checkbox"
@@ -120,8 +120,26 @@
                 <div class="modal-fullsettings-tab-expanded">
                     <div class="gh-stack">
                         <GhFormGroup class="gh-stack-item row">
+                            <label class="modal-fullsettings-title">Background color</label>
+                            <Modals::Newsletters::Components::ColorPicker
+                                @color={{@newsletter.backgroundColor}}
+                                @presetColors={{this.backgroundPresetColors}}
+                                @onColorChange={{fn this.changeSetting "backgroundColor"}}
+                            />
+                        </GhFormGroup>
+
+                        <GhFormGroup class="gh-stack-item row">
+                            <label class="modal-fullsettings-title">Border color</label>
+                            <Modals::Newsletters::Components::ColorPicker
+                                @color={{@newsletter.borderColor}}
+                                @presetColors={{this.borderPresetColors}}
+                                @onColorChange={{fn this.changeSetting "borderColor"}}
+                            />
+                        </GhFormGroup>
+
+                        <GhFormGroup class="gh-stack-item row">
                             <label class="modal-fullsettings-title">Post title</label>
-                            <div class="for-switch small">
+                            <div class="for-switch x-small">
                                 <label class="switch" for="show-post-title">
                                     <input
                                         type="checkbox"
@@ -137,6 +155,7 @@
 
                         {{#if @newsletter.showPostTitleSection}}
                             <GhFormGroup class="gh-stack-item">
+                                <label class="modal-fullsettings-title">Post title style</label>
                                 <div class="gh-email-design-typography-wrapper header">
                                     <div class="modal-fullsettings-radiogroup gh-email-design-typography" data-test-input="titleFontCategory">
                                         <GhFontSelector
@@ -152,8 +171,17 @@
                             </GhFormGroup>
                         {{/if}}
                         <GhFormGroup class="gh-stack-item row">
+                            <label class="modal-fullsettings-title">Heading color</label>
+                            <Modals::Newsletters::Components::ColorPicker
+                                @color={{@newsletter.titleColor}}
+                                @presetColors={{this.titlePresetColors}}
+                                @onColorChange={{fn this.changeSetting "titleColor"}}
+                            />
+                        </GhFormGroup>
+
+                        <GhFormGroup class="gh-stack-item row">
                             <label class="modal-fullsettings-title">Feature image</label>
-                            <div class="for-switch small">
+                            <div class="for-switch x-small">
                                 <label class="switch" for="show-feature-image">
                                     <input
                                         type="checkbox"
@@ -194,7 +222,7 @@
                         {{#if (feature "audienceFeedback")}}
                             <GhFormGroup class="gh-stack-item row">
                                 <label for="capture-feedback" class="modal-fullsettings-title" data-test-toggle="feedbackEnabled">Ask your readers for feedback</label>
-                                <div class="for-switch small">
+                                <div class="for-switch x-small">
                                     <div class="container">
                                         <input
                                             type="checkbox"
@@ -211,7 +239,7 @@
                         {{#unless (eq this.settings.commentsEnabled 'off') }}
                             <GhFormGroup class="gh-stack-item row">
                                 <label for="comments-link" class="modal-fullsettings-title" data-test-toggle="showCommentCta">Add a link to your comments</label>
-                                <div class="for-switch small">
+                                <div class="for-switch x-small">
                                     <div class="container">
                                         <input
                                             type="checkbox"
@@ -227,7 +255,7 @@
 
                         <GhFormGroup class="gh-stack-item row">
                             <label for="latest-posts" class="modal-fullsettings-title" data-test-toggle="showLatestPosts">Share your latest posts</label>
-                            <div class="for-switch small">
+                            <div class="for-switch x-small">
                                 <div class="container">
                                     <input
                                         type="checkbox"
@@ -242,7 +270,7 @@
 
                         <GhFormGroup class="gh-stack-item row">
                             <label for="capture-feedback" class="modal-fullsettings-title" data-test-toggle="showSubscriptionDetails">Show subscription details</label>
-                            <div class="for-switch small">
+                            <div class="for-switch x-small">
                                 <div class="container">
                                     <input
                                         type="checkbox"
