+++ ghost/email-service/lib/email-renderer.js
@@ -679,8 +679,13 @@ class EmailRenderer {
         });
 
         // Partials
-        const cssPartialSource = await fs.readFile(path.join(__dirname, './email-templates/partials/', `styles.hbs`), 'utf8');
-        this.#handlebars.registerPartial('styles', cssPartialSource);
+        if (this.#labs.isSet('makingItRain')) {
+            const cssPartialSource = await fs.readFile(path.join(__dirname, './email-templates/partials/', `styles.hbs`), 'utf8');
+            this.#handlebars.registerPartial('styles', cssPartialSource);
+        } else {
+            const cssPartialSource = await fs.readFile(path.join(__dirname, './email-templates/partials/', `styles-old.hbs`), 'utf8');
+            this.#handlebars.registerPartial('styles', cssPartialSource);
+        }
 
         const paywallPartial = await fs.readFile(path.join(__dirname, './email-templates/partials/', `paywall.hbs`), 'utf8');
         this.#handlebars.registerPartial('paywall', paywallPartial);
@@ -695,9 +700,13 @@ class EmailRenderer {
         this.#handlebars.registerPartial('latestPosts', latestPostsPartial);
 
         // Actual template
-        const htmlTemplateSource = await fs.readFile(path.join(__dirname, './email-templates/', `template.hbs`), 'utf8');
-        this.#renderTemplate = this.#handlebars.compile(Buffer.from(htmlTemplateSource).toString());
-
+        if (this.#labs.isSet('makingItRain')) {
+            const htmlTemplateSource = await fs.readFile(path.join(__dirname, './email-templates/', `template.hbs`), 'utf8');
+            this.#renderTemplate = this.#handlebars.compile(Buffer.from(htmlTemplateSource).toString());
+        } else {
+            const htmlTemplateSource = await fs.readFile(path.join(__dirname, './email-templates/', `template-old.hbs`), 'utf8');
+            this.#renderTemplate = this.#handlebars.compile(Buffer.from(htmlTemplateSource).toString());
+        }
         return this.#renderTemplate(data);
     }
